requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
DateTime==5.5
Django==5.2
django-braces==1.17.0
django-cors-headers==4.7.0
django-extensions==4.1
django-extra-views==0.8.0
django-filter==25.1
django-grappelli==4.0.2
django-jet==1.0.8
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
future==1.0.0
idna==3.10
logging==0.4.9.6
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
python-slugify==8.0.4
pytz==2016.4
redis==5.2.1
requests==2.32.3
setuptools==78.1.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
urllib3==2.3.0
zope.interface==7.2
